캐릭터 맵 안에 가두기 + 카메라 맵 안에 가두기

+++ ISAAC/APIClass/CButton.cpp
@@ -41,7 +41,7 @@ void CButton::render(HDC _dc)
 	}
 
 	Vec2 vPos = GetFinalPos();
-	BitBlt(_dc, vPos.x, vPos.y, pTargetTex->Widht(), pTargetTex->Height(), pTargetTex->GetDC(), 0, 0, SRCCOPY);
+	BitBlt(_dc, (int)vPos.x, (int)vPos.y, pTargetTex->Widht(), pTargetTex->Height(), pTargetTex->GetDC(), 0, 0, SRCCOPY);
